Добавлен модуль conftest.py в tests. Откорректированы функции. Откорректированы тесты. Фикстуры перенесены в conftest.py
