--- src/__init__.py
@@ -1,2 +1,0 @@
-def widgetimport():
-    return None
