+++ tests/conftest.py
@@ -1,12 +1,33 @@
 import pytest
 
-from src.masks import get_mask_account, get_mask_card_number
-from src.processing import filter_by_state, sort_by_date
-from src.widget import get_date, mask_account_card
+
+# Фикстуры для тестирования номера карты
+@pytest.fixture
+def correct_card():
+    return "1234567812345678"
 
 
-# Тестирование функции `get_mask_account`
-# Фикстура для тестирования
+@pytest.fixture
+def short_card():
+    return "12345678"  # 8 цифр
+
+
+@pytest.fixture
+def long_card():
+    return "12345678123456789"  # 17 цифр
+
+
+@pytest.fixture
+def invalid_card_chars():
+    return "1234abcd5678efgh"  # Содержит буквы
+
+
+@pytest.fixture
+def empty_card():
+    return ""  # Пустая строка
+
+
+# Фикстуры для тестирования номера счета
 @pytest.fixture
 def valid_account():
     return "1234567812345678"
@@ -32,77 +53,7 @@ def empty_account():
     return ""  # Пустая строка
 
 
-# Тесты с использованием фикстур
-def test_valid_account(valid_account):
-    assert get_mask_account(valid_account) == "**5678"
-
-
-def test_short_account(short_account):
-    assert get_mask_account(short_account) == "Ошибка: слишком короткий номер."
-
-
-def test_long_account(long_account):
-    assert get_mask_account(long_account) == "Ошибка: слишком длинный номер."
-
-
-def test_invalid_account_chars(invalid_account_chars):
-    assert get_mask_account(invalid_account_chars) == "Ошибка: присутствуют другие символы."
-
-
-def test_empty_account(empty_account):
-    assert get_mask_account(empty_account) == "Ошибка: входная строка пустая."
-
-
-# Тестирование функции `get_mask_card_number`
-# Фикстуры для тестирования
-@pytest.fixture
-def valid_card():
-    return "1234567812345678"
-
-
-@pytest.fixture
-def short_card():
-    return "12345678"  # 8 цифр
-
-
-@pytest.fixture
-def long_card():
-    return "12345678123456789"  # 17 цифр
-
-
-@pytest.fixture
-def invalid_card_chars():
-    return "1234abcd5678efgh"  # Содержит буквы
-
-
-@pytest.fixture
-def empty_card():
-    return ""  # Пустая строка
-
-
-# Тесты с использованием фикстур.
-def test_valid_card(valid_card):
-    assert get_mask_card_number(valid_card) == "1234 56** **** 5678"
-
-
-def test_short_card(short_card):
-    assert get_mask_card_number(short_card) == "Ошибка: слишком короткий номер."
-
-
-def test_long_card(long_card):
-    assert get_mask_card_number(long_card) == "Ошибка: слишком длинный номер."
-
-
-def test_invalid_card_chars(invalid_card_chars):
-    assert get_mask_card_number(invalid_card_chars) == "Ошибка: присутствуют другие символы."
-
-
-def test_empty_card(empty_card):
-    assert get_mask_card_number(empty_card) == "Ошибка: входная строка пустая."
-
-
-# Тестирование функции `filter_by_state`
-# Фикстуры для тестирования
+# Фикстуры для тестирования `state`
 @pytest.fixture
 def sample_data():
     return [
@@ -114,24 +65,7 @@ def sample_data():
     ]
 
 
-# Тесты с использованием фикстур.
-def test_empty_data():
-    assert filter_by_state([], "EXECUTED") == "Ошибка: входная строка пустая."
-
-
-def test_missing_state_key():
-    assert filter_by_state([{"id": 1}], "EXECUTED") == "Ошибка: такого ключа нет."
-
-
-def test_none_state_value():
-    assert (
-        filter_by_state([{"id": 1, "state": None}], "EXECUTED")
-        == "Ошибка: значение 'EXECUTED' не найдено для ключа 'state'."
-    )
-
-
-# Тестирование функции `sort_by_date`
-# Фикстуры для тестирования
+# Фикстуры для тестирования `date`
 @pytest.fixture
 def data_sample():
     return [
@@ -142,24 +76,7 @@ def data_sample():
     ]
 
 
-# Тесты с использованием фикстур.
-def test_empty_list():
-    with pytest.raises(ValueError, match="Ошибка: передан пустой список."):
-        sort_by_date([])
-
-
-def test_missing_date_key():
-    with pytest.raises(ValueError, match="Ошибка: отсутствует ключ 'date'"):
-        sort_by_date([{"not_date": "2023-01-01"}])
-
-
-def test_none_date_value():
-    with pytest.raises(ValueError, match="Ошибка: значение по ключу 'date'"):
-        sort_by_date([{"date": None}])
-
-
-# Тестирование функции mask_account_card
-# Фикстура для тестирования
+# Фикстуры для тестирования карт
 @pytest.fixture
 def valid_visa_card():
     return "Visa 1234567812345678"
@@ -171,27 +88,27 @@ def valid_mastercard():
 
 
 @pytest.fixture
-def valid_account_1():
+def correct_account():
     return "Счёт 1234567812345678"
 
 
 @pytest.fixture
-def short_card_1():
+def short_number_card():
     return "Visa 12345678"  # 8 цифр
 
 
 @pytest.fixture
-def long_card_1():
+def long_number_card():
     return "Visa 12345678123456789"  # 17 цифр
 
 
 @pytest.fixture
-def invalid_card_chars_1():
+def incorrect_card_chars():
     return "Visa 1234abcd5678efgh"  # Содержит буквы
 
 
 @pytest.fixture
-def empty_card_1():
+def card_no_data():
     return ""  # Пустая строка
 
 
@@ -200,42 +117,7 @@ def no_type_card():
     return "1234567812345678"  # Нет типа карты
 
 
-# Тесты с использованием фикстур
-def test_valid_visa_card(valid_visa_card):
-    assert mask_account_card(valid_visa_card) == "Visa 1234 56 ** **** 5678"
-
-
-def test_valid_mastercard(valid_mastercard):
-    assert mask_account_card(valid_mastercard) == "MasterCard 1234567812345678"
-
-
-def test_valid_account_1():
-    valid_account = "Счёт 1234567812345678"  # Передаем тип и номер
-    assert mask_account_card(valid_account) == "Счёт **5678"
-
-
-def test_short_card_1(short_card):
-    assert mask_account_card(short_card) == "Ошибка: неверно указан номер карты/счёта."
-
-
-def test_long_card_1(long_card):
-    assert mask_account_card(long_card) == "Ошибка: неверно указан номер карты/счёта."
-
-
-def test_invalid_card_chars_1(invalid_card_chars):
-    assert mask_account_card(invalid_card_chars) == "Ошибка: неверно указан номер карты/счёта."
-
-
-def test_empty_card_1(empty_card):
-    assert mask_account_card(empty_card) == "Ошибка: входная строка пустая."
-
-
-def test_no_type_card(no_type_card):
-    assert mask_account_card(no_type_card) == "Ошибка: номер указан без типа карты/счёта."
-
-
-# Тестирование функции `get_date`
-# Фикстуры для тестирования
+# Фикстуры для тестирования дат
 @pytest.fixture
 def valid_date():
     return "2023-03-15T12:30:00"  # Корректная дата
@@ -274,36 +156,3 @@ def invalid_day():
 @pytest.fixture
 def invalid_characters():
     return "2023-03-1aT12:30:00"  # Неверные символы в дате
-
-
-# Тесты с использованием фикстур
-def test_valid_date(valid_date):
-    assert get_date(valid_date) == "15.03.2023"
-
-
-def test_leap_year_date(leap_year_date):
-    assert get_date(leap_year_date) == "29.02.2020"
-
-
-def test_invalid_date_format(invalid_date_format):
-    assert get_date(invalid_date_format) == "Ошибка: неверный формат даты."
-
-
-def test_empty_date(empty_date):
-    assert get_date(empty_date) == "Ошибка: входная строка пустая."
-
-
-def test_no_time_part(no_time_part):
-    assert get_date(no_time_part) == "Ошибка: неверный формат даты."
-
-
-def test_invalid_month(invalid_month):
-    assert get_date(invalid_month) == "Ошибка: некорректный месяц."
-
-
-def test_invalid_day(invalid_day):
-    assert get_date(invalid_day) == "Ошибка: некорректный день."
-
-
-def test_invalid_characters(invalid_characters):
-    assert get_date(invalid_characters) == "Ошибка: дата содержит недопустимые символы."
